added cloudinary

+++ RSVP.Application/Features/Images/Commands/DeleteImagesInGroup/DeleteImagessInGroupCommandHandler.cs
@@ -5,20 +5,22 @@ using appDomain = RSVP.Domain.Entities;
 
 namespace RSVP.Application.Features.Images.Commands.DeleteImagesInGroup;
 
-public class DeleteImagessInGroupCommandHandler:IRequestHandler<DeleteImagessInGroupCommand,bool>
+public class DeleteImagessInGroupCommandHandler:IRequestHandler<DeleteImagesInGroupCommand,bool>
 {
     private readonly IRepository<appDomain.Media> _mediarepository;
     private readonly IEventAccessService _eventAccessService;
     private readonly IRsvpDbContext _dbContext;
-    public DeleteImagessInGroupCommandHandler(IRepository<appDomain.Media> mediarepository, IEventAccessService eventAccessService, IRsvpDbContext dbContext)
+    private readonly IPhotoService _photoService;
+    public DeleteImagessInGroupCommandHandler(IRepository<appDomain.Media> mediarepository, IEventAccessService eventAccessService, IRsvpDbContext dbContext, IPhotoService photoService)
     {
         _mediarepository = mediarepository;
         _eventAccessService = eventAccessService;
         _dbContext = dbContext;
+        _photoService = photoService;
     }
-    public async Task<bool> Handle(DeleteImagessInGroupCommand request, CancellationToken cancellationToken)
+    public async Task<bool> Handle(DeleteImagesInGroupCommand request, CancellationToken cancellationToken)
     {
-       List<appDomain.Media> imagesToDelete = _dbContext.Media.Where(m => request.ImageIds.Contains(m.Id)).ToList();
+       List<appDomain.Media> imagesToDelete = _dbContext.Media.Where(m => request.ImageIds.Select(i => i.ImageId).Contains(m.Id)).ToList();
 
         if (imagesToDelete.Count == 0)
         {
@@ -39,6 +41,11 @@ public class DeleteImagessInGroupCommandHandler:IRequestHandler<DeleteImagessInG
         if (!isOrganizerOrOwner)
         {
             throw new UnauthorizedAccessException("User is not authorized to delete images from this event.");
+        }
+        long count = await _photoService.DeleteMultiplePhotosAsync(request.ImageIds.Select(i => i.PublicId).ToList());
+        if(count == 0)
+        {
+            throw new Exception("Failed to delete images from the photo service.");
         }
 
         _mediarepository.DeleteRange(imagesToDelete);
